Rename dynamic-macros.h to dynamic_macros.h
