+++ app/tests/dynamic-macros/behavior_keymap.dtsi
@@ -7,7 +7,7 @@
 #include <dt-bindings/zmk/keys.h>
 #include <behaviors.dtsi>
 #include <dt-bindings/zmk/kscan_mock.h>
-#include <dt-bindings/zmk/dynamic-macros.h>
+#include <dt-bindings/zmk/dynamic_macros.h>
 
 / {
     macros {
